v1.1 - FU
In this versions, the modifications done aimed correcting redundacies
created when trying to fix some warnings (there was a warning that
mentioned that a txt file was unclosed, even using "with". In the end,
the error was in other line...). Also, instructions were written about
the paths and some clean in the code was also performed.