v1.2 - FU
In this version, an extension was added. The table joiner was done to
join the files converted to csv. As a first release, it only joins the
table with (almost) no filtering. Some notes were done troughout the
code about errors and warnings risen. Also, you may check the pandas
functionality "concat" to understand more about the concatenations type.
In my use, the axis was assigned as 1, solving the desired problem.

Some error may be risen when converting the csv file. It is caused by "," in the file.
Error message--> pandas.errors.ParserError: Error	tokenizing data. C error: Expected...
To handle with it, manually, remove the characters that are breaking the code (Ctrl-F solves).
If you may want, the program can pass through it and display what files had problems.
To do so, simply remove the """ below and put the line 27 inside the try.
Other errors may appear to due to the enconding. You can change the encode argument of read_csv(utf-8).