v0.2
In this version, the insertion of initialization and forward propagation functions were done.
Both of them will be updated in future releases to better suit in the
project. About the forward propagation function, there are 3 main
funcitons: neuron_activation, which activates the neuron, ReLU, the
activation function of the neuron, and forward_propagation, propagates
the weights all over the ANN in direction to the output.