v0.3
In this update some corrections were done to fix bugs and redudancies in the
code. Also, as said before, the adaptation of the code to MSP430 family was started.
It is worth mentioning that in this project a MSP430F5529 is being used. In addition,
another function was added aiming to compute the output class, since the
ANN is for classification.

