v0.4
In this version, some corrections were done in the ANN structure to
allow the correct forward propagation of the results. In addition, some
functions were delete since they were no longer necessary. Moreover,
aiming the implementation in microcontrollers, such as MSP430, all the
allocation of memory was done statically to avoid problem regarding
overlaping memory of the stack or other variables. Having this in hand,
test will be done to ensure that the propagation is performing
correctly.