v0.5
In this update, the test of the forward propagation function was done.
Some mistakes were found during the correction and are alrefy fixed.
Also, adaptations were done to enable MSP430F5529 reading the file and
running it. The code was also tested in MSP430, and positive results were seen.
The next step is creating a function to classify the sample according
to the results given by the network.