v0.6
In this update, some corrections were done in the propagation. Given
that the ANN should be used more than once, it was changed the variables in which the results were stored.
Moreover, some test regarding memory overflow were done. The worst
case scenario was set as all inputs and bias being 1023, and the weights being 128.
The result of it gives a maximum value which represent 93.75% of the size
supported by a long data type.