v0.7
In this version a big fix in the code was done regarding the structure
of the ANN (the multiplication between the layers was wrong). Moreover,
the problem of overflow in the previous version was overcome because
in simulations done, it was seen that on worst case scenarios for the input, the
overflow did not occurred because the negative weights helped decreasing
the values of the sums. Also, if necessary, more allocation of
memory can be done (meaning that the layers can assume long
long data type) to solve the overflow issue and still have empty space to use.
In addition, the weights and biases were converted from float to char
to maintain low-usage of memory. However, it still can be converted to
int type witout problems regarding memory mapping.