v0.8
The changes concerning this update fixed various aspects of the program.
Firstly, it was corrected various overflow issues. It was thought that
no overflow was occurring because, after the multiplicaiton, big values
were appearing, and also no negative numbers were being seen what,
theoretically, is correct given that all negative values become 0 after
passing the activation function. However, the values were too big and
were caused by the overflow of the numbers supported by each data type.
Moreover, the zeros seen were also a result of the overeflow derived
from the way that the numbers were being multiplied. Despite the
significant increase in memory usage (after the correction to long long),
a lot of it reamins empty. Secondly, modifications in the data type of
the weights and biases were done because, in MSP430, the multiplicaitons
between values of different data types does not behave in the same way
that happens in the "standard" C (which was beforehand used to test the
code aiming to the better understanding of erros). In MSP430 code, every
data type operation have to be explicity to work properly, making it
necessaty the usage of typecast. Finally, it was checked if the forward
propagation was correctly working, and the result was affirmative.